+++ src/test/unit/EigenPodUnit.t.sol
@@ -56,7 +56,6 @@ contract EigenPodUnitTests is EigenLayerUnitTestSetup, EigenPodPausingConstants,
         timeMachine = new TimeMachine();
         beaconChain = new BeaconChainMock(EigenPodManager(address(eigenPodManagerMock)), GENESIS_TIME_LOCAL);
 
-
         // Deploy EigenPod
         podImplementation = new EigenPod(ethPOSDepositMock, IEigenPodManager(address(eigenPodManagerMock)), GENESIS_TIME_LOCAL, "v9.9.9");
 
@@ -115,7 +114,7 @@ contract EigenPodUnitTests is EigenLayerUnitTestSetup, EigenPodPausingConstants,
         stakerName = string.concat("Staker", cheats.toString(numStakers));
         staker = new EigenPodUser(stakerName);
 
-        uint256 amount = bound(rand, 1 ether, 4096 ether);
+        uint amount = bound(rand, 1 ether, 4096 ether);
         cheats.deal(address(staker), amount);
 
         numStakers++;
@@ -633,7 +632,7 @@ contract EigenPodUnitTests_verifyWithdrawalCredentials is EigenPodUnitTests, Pro
 
     /// @notice test verify wc reverts when paused
     function test_revert_verifyWithdrawalCredentialsPaused() public {
-        (EigenPodUser staker,) = _newEigenPodStaker({ rand: 0 });
+        (EigenPodUser staker,) = _newEigenPodStaker({rand: 0});
         (uint40[] memory validators,,) = staker.startValidators();
 
         cheats.prank(pauser);
@@ -646,7 +645,7 @@ contract EigenPodUnitTests_verifyWithdrawalCredentials is EigenPodUnitTests, Pro
     /// @notice beaconTimestamp must be after the current checkpoint
     function testFuzz_revert_beaconTimestampInvalid() public {
         cheats.warp(10 days);
-        (EigenPodUser staker,) = _newEigenPodStaker({ rand: 0 });
+        (EigenPodUser staker,) = _newEigenPodStaker({rand: 0});
         // Ensure we have more than one validator (_newEigenPodStaker allocates a nonzero amt of eth)
         cheats.deal(address(staker), address(staker).balance + 32 ether);
         (uint40[] memory validators,,) = staker.startValidators();
@@ -664,8 +663,8 @@ contract EigenPodUnitTests_verifyWithdrawalCredentials is EigenPodUnitTests, Pro
     }
 
     /// @notice Check for revert on input array mismatch lengths
-    function testFuzz_revert_inputArrayLengthsMismatch(uint256 rand) public {
-        (EigenPodUser staker,) = _newEigenPodStaker({ rand: rand });
+    function testFuzz_revert_inputArrayLengthsMismatch(uint rand) public {
+        (EigenPodUser staker,) = _newEigenPodStaker({rand: rand});
         (uint40[] memory validators,,) = staker.startValidators();
         EigenPod pod = staker.pod();
         CredentialProofs memory proofs = beaconChain.getCredentialProofs(validators);
@@ -706,8 +705,8 @@ contract EigenPodUnitTests_verifyWithdrawalCredentials is EigenPodUnitTests, Pro
     }
 
     /// @notice Check beaconStateRootProof reverts on invalid length or invalid proof
-    function testFuzz_revert_beaconStateRootProofInvalid(uint256 rand) public {
-        (EigenPodUser staker,) = _newEigenPodStaker({ rand: rand });
+    function testFuzz_revert_beaconStateRootProofInvalid(uint rand) public {
+        (EigenPodUser staker,) = _newEigenPodStaker({rand: rand});
         (uint40[] memory validators,,) = staker.startValidators();
         EigenPod pod = staker.pod();
         CredentialProofs memory proofs = beaconChain.getCredentialProofs(validators);
@@ -744,8 +743,8 @@ contract EigenPodUnitTests_verifyWithdrawalCredentials is EigenPodUnitTests, Pro
 
     /// @notice attempt to verify validator credentials in both ACTIVE and WITHDRAWN states
     /// check reverts
-    function testFuzz_revert_validatorsWithdrawn(uint256 rand) public {
-        (EigenPodUser staker,) = _newEigenPodStaker({ rand: rand });
+    function testFuzz_revert_validatorsWithdrawn(uint rand) public {
+        (EigenPodUser staker,) = _newEigenPodStaker({rand: rand});
         (uint40[] memory validators,,) = staker.startValidators();
         staker.verifyWithdrawalCredentials(validators);
 
@@ -767,10 +766,10 @@ contract EigenPodUnitTests_verifyWithdrawalCredentials is EigenPodUnitTests, Pro
 
     /// @notice attempt to verify validator credentials after they have exited
     /// check reverts
-    function testFuzz_revert_validatorsExited(uint256 rand) public {
-        (EigenPodUser staker,) = _newEigenPodStaker({ rand: rand });
+    function testFuzz_revert_validatorsExited(uint rand) public {
+        (EigenPodUser staker,) = _newEigenPodStaker({rand: rand});
         (uint40[] memory validators,,) = staker.startValidators();
-        
+
         // Exit validators from beacon chain and withdraw to pod
         staker.exitValidators(validators);
         beaconChain.advanceEpoch();
@@ -781,8 +780,8 @@ contract EigenPodUnitTests_verifyWithdrawalCredentials is EigenPodUnitTests, Pro
     }
 
     /// @notice modify withdrawal credentials to cause a revert
-    function testFuzz_revert_invalidWithdrawalAddress(uint256 rand, bytes32 invalidWithdrawalCredentials) public {
-        (EigenPodUser staker,) = _newEigenPodStaker({ rand: rand });
+    function testFuzz_revert_invalidWithdrawalAddress(uint rand, bytes32 invalidWithdrawalCredentials) public {
+        (EigenPodUser staker,) = _newEigenPodStaker({rand: rand});
         (uint40[] memory validators,,) = staker.startValidators();
         EigenPod pod = staker.pod();
         CredentialProofs memory proofs = beaconChain.getCredentialProofs(validators);
@@ -803,10 +802,9 @@ contract EigenPodUnitTests_verifyWithdrawalCredentials is EigenPodUnitTests, Pro
         cheats.stopPrank();
     }
 
-
     /// @notice test verify wc reverts when fork timestamp is zero
     function test_revert_forkTimestampZero() public {
-        (EigenPodUser staker,) = _newEigenPodStaker({ rand: 0 });
+        (EigenPodUser staker,) = _newEigenPodStaker({rand: 0});
         (uint40[] memory validators,,) = staker.startValidators();
 
         // Set zero fork timestamp to zero
@@ -817,8 +815,8 @@ contract EigenPodUnitTests_verifyWithdrawalCredentials is EigenPodUnitTests, Pro
     }
 
     /// @notice modify validator field length to cause a revert
-    function testFuzz_revert_invalidValidatorFields(uint256 rand) public {
-        (EigenPodUser staker,) = _newEigenPodStaker({ rand: rand });
+    function testFuzz_revert_invalidValidatorFields(uint rand) public {
+        (EigenPodUser staker,) = _newEigenPodStaker({rand: rand});
         (uint40[] memory validators,,) = staker.startValidators();
         EigenPod pod = staker.pod();
         CredentialProofs memory proofs = beaconChain.getCredentialProofs(validators);
@@ -843,8 +841,8 @@ contract EigenPodUnitTests_verifyWithdrawalCredentials is EigenPodUnitTests, Pro
     }
 
     /// @notice modify validator activation epoch to cause a revert
-    function testFuzz_revert_activationEpochNotSet(uint256 rand) public {
-        (EigenPodUser staker,) = _newEigenPodStaker({ rand: rand });
+    function testFuzz_revert_activationEpochNotSet(uint rand) public {
+        (EigenPodUser staker,) = _newEigenPodStaker({rand: rand});
         (uint40[] memory validators,,) = staker.startValidators();
         EigenPod pod = staker.pod();
         CredentialProofs memory proofs = beaconChain.getCredentialProofs(validators);
@@ -865,8 +863,8 @@ contract EigenPodUnitTests_verifyWithdrawalCredentials is EigenPodUnitTests, Pro
     }
 
     /// @notice modify validator proof length to cause a revert
-    function testFuzz_revert_invalidValidatorProofLength(uint256 rand) public {
-        (EigenPodUser staker,) = _newEigenPodStaker({ rand: rand });
+    function testFuzz_revert_invalidValidatorProofLength(uint rand) public {
+        (EigenPodUser staker,) = _newEigenPodStaker({rand: rand});
         (uint40[] memory validators,,) = staker.startValidators();
         EigenPod pod = staker.pod();
         CredentialProofs memory proofs = beaconChain.getCredentialProofs(validators);
@@ -887,7 +885,7 @@ contract EigenPodUnitTests_verifyWithdrawalCredentials is EigenPodUnitTests, Pro
     }
 
     /// @notice modify validator pubkey to cause a revert
-    function testFuzz_revert_invalidValidatorProof(uint256 rand, bytes32 randPubkey) public {
+    function testFuzz_revert_invalidValidatorProof(uint rand, bytes32 randPubkey) public {
         (EigenPodUser staker,) = _newEigenPodStaker({rand: rand});
         (uint40[] memory validators,,) = staker.startValidators();
         EigenPod pod = staker.pod();
@@ -1076,10 +1074,7 @@ contract EigenPodUnitTests_verifyCheckpointProofs is EigenPodUnitTests {
         (EigenPodUser staker,) = _newEigenPodStaker({rand: rand});
         EigenPod pod = staker.pod();
         (uint40[] memory validators,,) = staker.startValidators();
-        CheckpointProofs memory proofs = beaconChain.getCheckpointProofs(
-            validators,
-            pod.currentCheckpointTimestamp()
-        );
+        CheckpointProofs memory proofs = beaconChain.getCheckpointProofs(validators, pod.currentCheckpointTimestamp());
 
         cheats.prank(pauser);
         eigenPodManagerMock.pause(2 ** PAUSED_EIGENPODS_VERIFY_CHECKPOINT_PROOFS);
@@ -1102,7 +1097,7 @@ contract EigenPodUnitTests_verifyCheckpointProofs is EigenPodUnitTests {
     /// @notice should revert if fork timestamp zero
     function test_revert_forkTimestampZero() public {
         // Setup verifyCheckpointProofs
-        (EigenPodUser staker,) = _newEigenPodStaker({ rand: 0 });
+        (EigenPodUser staker,) = _newEigenPodStaker({rand: 0});
         (uint40[] memory validators,,) = staker.startValidators();
         staker.verifyWithdrawalCredentials(validators);
         beaconChain.advanceEpoch();
@@ -1192,7 +1187,7 @@ contract EigenPodUnitTests_verifyCheckpointProofs is EigenPodUnitTests {
     /// @notice test that verifyCheckpointProofs skips proofs submitted for non-ACTIVE validators
     function testFuzz_verifyCheckpointProofs_skipIfNotActive(uint rand) public {
         // Setup verifyCheckpointProofs
-        (EigenPodUser staker,) = _newEigenPodStaker({ rand: rand });
+        (EigenPodUser staker,) = _newEigenPodStaker({rand: rand});
         EigenPod pod = staker.pod();
 
         (uint40[] memory validators1,,) = staker.startValidators();
@@ -1254,7 +1249,7 @@ contract EigenPodUnitTests_verifyCheckpointProofs is EigenPodUnitTests {
     /// @notice test that verifyCheckpointProofs skips duplicate checkpoint proofs
     function testFuzz_verifyCheckpointProofs_skipIfAlreadyProven(uint rand) public {
         // Setup verifyCheckpointProofs
-        (EigenPodUser staker,) = _newEigenPodStaker({ rand: rand });
+        (EigenPodUser staker,) = _newEigenPodStaker({rand: rand});
         EigenPod pod = staker.pod();
 
         (uint40[] memory validators1,,) = staker.startValidators();
@@ -1484,7 +1479,7 @@ contract EigenPodUnitTests_verifyStaleBalance is EigenPodUnitTests {
     /// @notice verifyStaleBalance should revert with forkTimestamp of 0
     function test_revert_forkTimestampZero() public {
         // setup eigenpod staker and validators
-        (EigenPodUser staker,) = _newEigenPodStaker({ rand: 0 });
+        (EigenPodUser staker,) = _newEigenPodStaker({rand: 0});
         EigenPod pod = staker.pod();
         (uint40[] memory validators,,) = staker.startValidators();
         uint40 validator = validators[0];
@@ -1506,8 +1501,8 @@ contract EigenPodUnitTests_verifyStaleBalance is EigenPodUnitTests {
         });
     }
 
-     /// @notice verifyStaleBalance should revert with invalid beaconStateRoot proof length
-    function testFuzz_revert_beaconStateRootProofInvalidLength(uint256 rand) public {
+    /// @notice verifyStaleBalance should revert with invalid beaconStateRoot proof length
+    function testFuzz_revert_beaconStateRootProofInvalidLength(uint rand) public {
         // setup eigenpod staker and validators
         (EigenPodUser staker,) = _newEigenPodStaker({rand: rand});
         EigenPod pod = staker.pod();
@@ -1654,20 +1649,17 @@ contract EigenPodUnitTests_verifyStaleBalance is EigenPodUnitTests {
 
 /// @notice Tests that deneb proofs are not successful against pectra state
 contract EigenPodUnitTests_DenebProofsAgainstPectra is EigenPodUnitTests {
-
     /// @notice Set the beacon chain mock to Deneb Forkable
     /// @notice The EigenPods by default use Pectra state given the early `proofTimestamp`
     function setUp() public override {
         EigenPodUnitTests.setUp();
 
         // Set beaconChainMock to Deneb Forkable
-        beaconChain = BeaconChainMock(
-            new BeaconChainMock_DenebForkable(EigenPodManager(address(eigenPodManagerMock)), GENESIS_TIME_LOCAL)
-        );
+        beaconChain = BeaconChainMock(new BeaconChainMock_DenebForkable(EigenPodManager(address(eigenPodManagerMock)), GENESIS_TIME_LOCAL));
     }
 
     function testFuzz_revert_verifyWC_DenebAgainstPectra(uint24 rand) public {
-        (EigenPodUser staker,) = _newEigenPodStaker({ rand: rand });
+        (EigenPodUser staker,) = _newEigenPodStaker({rand: rand});
         (uint40[] memory validators,,) = staker.startValidators();
         EigenPod pod = staker.pod();
 
@@ -1690,7 +1682,7 @@ contract EigenPodUnitTests_DenebProofsAgainstPectra is EigenPodUnitTests {
         eigenPodManagerMock.setPectraForkTimestamp(type(uint64).max);
 
         // Setup verifyCheckpointProofs
-        (EigenPodUser staker,) = _newEigenPodStaker({ rand: rand });
+        (EigenPodUser staker,) = _newEigenPodStaker({rand: rand});
         EigenPod pod = staker.pod();
         (uint40[] memory validators,,) = staker.startValidators();
         staker.verifyWithdrawalCredentials(validators);
@@ -1698,19 +1690,13 @@ contract EigenPodUnitTests_DenebProofsAgainstPectra is EigenPodUnitTests {
 
         // Get checkpoint proofs on Deneb state
         staker.startCheckpoint();
-        CheckpointProofs memory proofs = beaconChain.getCheckpointProofs(
-            validators,
-            pod.currentCheckpointTimestamp()
-        );
+        CheckpointProofs memory proofs = beaconChain.getCheckpointProofs(validators, pod.currentCheckpointTimestamp());
 
         // Rewind proof timestamp such that EigenPods prove everything against Pectra state
         eigenPodManagerMock.setPectraForkTimestamp(1 hours * 12);
 
         cheats.expectRevert(BeaconChainProofs.InvalidProofLength.selector);
-        pod.verifyCheckpointProofs({
-            balanceContainerProof: proofs.balanceContainerProof,
-            proofs: proofs.balanceProofs
-        });
+        pod.verifyCheckpointProofs({balanceContainerProof: proofs.balanceContainerProof, proofs: proofs.balanceProofs});
     }
 }
 
@@ -1779,13 +1765,7 @@ contract EigenPodUnitTests_proofParsingTests is EigenPodHarnessSetup, ProofParsi
         cheats.warp(GENESIS_TIME_LOCAL + 1 days);
         oracleTimestamp = uint64(block.timestamp);
 
-        eigenPodHarness.verifyWithdrawalCredentials(
-            oracleTimestamp,
-            beaconStateRoot,
-            validatorIndex,
-            validatorFieldsProof,
-            validatorFields
-        );
+        eigenPodHarness.verifyWithdrawalCredentials(oracleTimestamp, beaconStateRoot, validatorIndex, validatorFieldsProof, validatorFields);
         _;
     }
 }
